requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.14.0
apache-airflow-providers-slack==8.9.0
astronomer-cosmos==1.7.0
Faker==28.0.0